% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpdpgrow.R
\name{predict_functions}
\alias{predict_functions}
\title{Use the model-estimated covariance parameters from gpdpgrow() or gmrdpgrow to predict the function at
future time points.}
\usage{
predict_functions(object, J, ...)
}
\arguments{
\item{object}{Object of class \code{gpdpgrow} or\code{gmrfdpgrow()}.}

\item{J}{Scalar denoting number of draws to take from posterior predictive for each unit.
Defaults to \code{J = 500}.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
This is the generic predict_functions method. See the following functions
for the details about different data structures:
}
\details{
\itemize{
  \item \code{\link{predict_functions.gpdpgrow}} for objects of class \code{gpdpgrow}
  \item \code{\link{predict_functions.gmrfdpgrow}} for objects of class \code{gmrfdpgrow}
}
}
\keyword{manip}
