% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpdpgrow.R
\name{samples}
\alias{samples}
\alias{samples.gpdpgrow}
\title{Produce samples of MCMC output}
\usage{
samples(object, ...)
}
\arguments{
\item{object}{A \code{gpdpgrow} object}

\item{...}{Ignored}
}
\description{
provides posterior sampled values for every model parameter of a
\code{gpdpgrow} object
}
