% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmrfdpgrow.R
\name{samples.gmrfdpgrow}
\alias{samples.gmrfdpgrow}
\title{Produce samples of MCMC output}
\usage{
\method{samples}{gmrfdpgrow}(object, ...)
}
\arguments{
\item{object}{A \code{gmrfdpgrow} object}

\item{...}{Ignored}
}
\description{
provides posterior sampled values for every model parameter of a
\code{gmrfdpgrow} object
}
